+++ pkgs/data_provider/lib/data_provider.dart
@@ -1,5 +1,6 @@
 export 'src/exception.dart';
 export 'src/operator.dart';
+export 'src/projector.dart';
 export 'src/serializable.dart';
 export 'src/sorter.dart';
 export 'src/specification.dart';