+++ pkgs/data_provider/lib/src/data_provider.dart
@@ -1,6 +1,5 @@
-import 'package:data_provider/data_provider.dart';
-import 'package:data_provider/src/projector.dart';
-
+import 'operator.dart';
+import 'projector.dart';
 import 'sorter.dart';
 import 'specification.dart';